Added example.py & improved_table.xlsx
A simple example of some improvements that might improve your workflow on the analysis. Reading in the data is quite an expensive (slow) operation, so doing it once instead of every time speeds up the performance notably so you don't have to spend quite as long waiting for it to run.